a fix here. a fix there. implemented partly some of the filerequester's sub requesters.
git-svn-id: https://svn.aros.org/svn/aros/trunk@9095 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/asl/modereqsupport.c
@@ -665,7 +665,7 @@ void SMRefreshPropertyWindow(struct LayoutData *ld, struct DisplayMode *dispmode
     };
     struct Node 	*node = udata->PropertyNodes;
 
-    #define OF(x) (OFFSET(IntSMReq, ism_PropertyList_ ## x))
+    #define OF(x) (ULONG)(OFFSET(IntSMReq, ism_PropertyList_ ## x))
     
     static struct propertyinfo
     {
